refactor(cqrs): rename history.ts into cqrs/undo/
Move src/core/store/history.ts -> src/core/cqrs/undo/historyStore.ts.
Selector contract preserved (canUndo/canRedo/undo/redo/clear/push/
captureSelectionSnapshot/SelectionSnapshot/HistoryEntry). All caller
files updated to import from the new path.

Why undo/ belongs under cqrs/, not core/store/:
- Snapshots are pushed by undoCaptureMiddleware (cqrs/middleware/)
- Toasts read commandDescriptions (cqrs/commandDescriptions.ts) for the
  human-facing action label
- The store imports CommandType from cqrs/commands

Keeping the file under core/store/ created a directional cycle:
cqrs/middleware/undoCapture -> core/store/history -> cqrs/commands.
After this move, the cycle is gone — undo lives entirely on the cqrs
side, with one-way deps into core/store/{layout,selection,toast}.

Pure rename. Behavioral parity preserved across the entire selector
surface; tests verify (1301 tests, 89 files in changed area).
